Split Traverser.kt into Traverser.kt and UtBotSymbolicEngine.kt
Convert Traversal functions to extension functions
